refactor: Extensions splited into `ErrorExtensions` and `ResultExtensions`
